Bug #1 - Utilization with 10GbE nics on RHEL is incorrect
This fixes http://sourceforge.net/p/nicstat/bugs/1/ by applying the
patch provided by Darren in that that bug report.

This bug occurs due to an integer overflow for 10Gb interfaces and
is not specific to RHEL.

+++ nicstat.c
@@ -1609,7 +1609,7 @@ get_speed_duplex(nicdata_t *nicp)
 		get_speed_duplex(nicp);
 		return;
 	}
-	nicp->speed = edata.speed * 1000000;
+	nicp->speed = (long long) edata.speed * 1000000;
 	nicp->duplex = edata.duplex;
 }
 #endif /* OS_LINUX */
